RealTime Suggestions on SearchPage
Passed fuse object as a prop to the on change handler of the input search box and limited to 5 results. Need to implement for home page search bar too. Would probably be best to save the data api json read upon entering of website and the search function that would come with it in the state and pass that prop down